feat(ui,glue): interface, glue and client functions from the reference (wave 0)
Level-0 functions from their decompilations, tagged, with the Lua bindings that became portable:
the arena team bindings, GetGMTicketCategories, GetTitleName, GetObjectiveText, GetProgressText,
GetRewardSpell, GetMultiCastTotemSpells, SetRaidRosterSelection and three calendar event
bindings; the calendar, guild bank, auction, currency, reputation, equipment-set, talent, socket,
quest-frame and friend-list state they read (new zero-initialised globals tagged with the
reference addresses); party member records, action bar events, camera setters, cursor
accessors, mirror timers and totem slots, character name validation, quest text variables,
item stat totals for tooltips, input control, CLayoutFrame protection hooks, CSimpleTop and
CScriptRegion helpers. Structs whose full layout is not recovered declare only the offsets the
decompilations read.

Compiled and linked on the Linux build; not run.

+++ src/net/grunt/ClientLink.cpp
@@ -645,7 +645,7 @@ void Grunt::ClientLink::PackLogon(CDataStore& msg, const Logon& logon) {
     msg.Put(accountNameLen);
     msg.PutData(this->m_accountName, accountNameLen);
 
-    msg.Set(startPos, msg.Size() - startPos - 2);
+    msg.Set(startPos, static_cast<uint16_t>(msg.Size() - startPos - 2));
 }
 
 void Grunt::ClientLink::ProveVersion(const uint8_t* versionChecksum) {
